+++ sockeye/arguments.py
@@ -483,14 +483,18 @@ def add_device_args(params):
 def add_vocab_args(params):
     params.add_argument('--source-vocab',
                         required=False,
-                        nargs='+',
-                        type=regular_file(),
-                        default=[None],
-                        help='Existing source vocabulary (-ies) (JSON).')
+                        default=None,
+                        help='Existing source vocabulary (JSON).')
     params.add_argument('--target-vocab',
                         required=False,
                         default=None,
                         help='Existing target vocabulary (JSON).')
+    params.add_argument('--source-factor-vocabs',
+                        required=False,
+                        nargs='+',
+                        type=regular_file(),
+                        default=[],
+                        help='Existing source factor vocabulary (-ies) (JSON).')
     params.add_argument(C.VOCAB_ARG_SHARED_VOCAB,
                         action='store_true',
                         default=False,